+++ tests/sequential/prometheus_metrics/prometheus_metrics_test.go
@@ -26,6 +26,7 @@ import (
 const (
 	testName              = "prometheus-metrics-test"
 	labelScaledObject     = "scaledObject"
+	labelScaledJob        = "scaledJob"
 	labelType             = "type"
 	labelCloudEventSource = "cloudeventsource"
 	eventsink             = "eventsink"
@@ -39,7 +40,9 @@ var (
 	deploymentName                 = fmt.Sprintf("%s-deployment", testName)
 	monitoredDeploymentName        = fmt.Sprintf("%s-monitored", testName)
 	scaledObjectName               = fmt.Sprintf("%s-so", testName)
-	wrongScaledObjectName          = fmt.Sprintf("%s-wrong", testName)
+	wrongScaledObjectName          = fmt.Sprintf("%s-so-wrong", testName)
+	scaledJobName                  = fmt.Sprintf("%s-sj", testName)
+	wrongScaledJobName             = fmt.Sprintf("%s-sj-wrong", testName)
 	wrongScalerName                = fmt.Sprintf("%s-wrong-scaler", testName)
 	cronScaledJobName              = fmt.Sprintf("%s-cron-sj", testName)
 	clientName                     = fmt.Sprintf("%s-client", testName)
@@ -59,7 +62,9 @@ type templateData struct {
 	TestNamespace              string
 	DeploymentName             string
 	ScaledObjectName           string
+	ScaledJobName              string
 	WrongScaledObjectName      string
+	WrongScaledJobName         string
 	WrongScalerName            string
 	CronScaledJobName          string
 	MonitoredDeploymentName    string
@@ -163,6 +168,68 @@ spec:
         query: 'keda_scaler_errors_total{namespace="{{.TestNamespace}}",scaledObject="{{.WrongScaledObjectName}}"}'
 `
 
+	scaledJobTemplate = `
+apiVersion: keda.sh/v1alpha1
+kind: ScaledJob
+metadata:
+  name: {{.ScaledJobName}}
+  namespace: {{.TestNamespace}}
+spec:
+  jobTargetRef:
+    template:
+      spec:
+        containers:
+        - name: external-executor
+          image: busybox
+          command:
+          - sleep
+          - "30"
+          imagePullPolicy: IfNotPresent
+        restartPolicy: Never
+    backoffLimit: 1
+  pollingInterval: 5
+  maxReplicaCount: 3
+  successfulJobsHistoryLimit: 0
+  failedJobsHistoryLimit: 0
+  triggers:
+    - type: kubernetes-workload
+      metadata:
+        podSelector: 'app={{.MonitoredDeploymentName}}'
+        value: '1'
+`
+
+	wrongScaledJobTemplate = `
+apiVersion: keda.sh/v1alpha1
+kind: ScaledJob
+metadata:
+  name: {{.WrongScaledJobName}}
+  namespace: {{.TestNamespace}}
+spec:
+  jobTargetRef:
+    template:
+      spec:
+        containers:
+        - name: external-executor
+          image: busybox
+          command:
+          - sleep
+          - "30"
+          imagePullPolicy: IfNotPresent
+        restartPolicy: Never
+    backoffLimit: 1
+  pollingInterval: 2
+  maxReplicaCount: 3
+  successfulJobsHistoryLimit: 0
+  failedJobsHistoryLimit: 0
+  triggers:
+    - type: prometheus
+      name: {{.WrongScalerName}}
+      metadata:
+        serverAddress: http://keda-prometheus.keda.svc.cluster.local:8080
+        metricName: keda_scaler_errors_total
+        threshold: '1'
+        query: 'keda_scaler_errors_total{namespace="{{.TestNamespace}}",scaledJob="{{.WrongScaledJobName}}"}'
+`
 	cronScaledJobTemplate = `
 apiVersion: keda.sh/v1alpha1
 kind: ScaledJob
@@ -374,6 +441,7 @@ func TestPrometheusMetrics(t *testing.T) {
 	testScalerMetricLatency(t)
 	testScalerActiveMetric(t)
 	testScaledObjectErrors(t, data)
+	testScaledJobErrors(t, data)
 	testScalerErrors(t, data)
 	testScalerErrorsTotal(t, data)
 	testOperatorMetrics(t, kc, data)
@@ -394,6 +462,8 @@ func getTemplateData() (templateData, []Template) {
 			DeploymentName:             deploymentName,
 			ScaledObjectName:           scaledObjectName,
 			WrongScaledObjectName:      wrongScaledObjectName,
+			ScaledJobName:              scaledJobName,
+			WrongScaledJobName:         wrongScaledJobName,
 			WrongScalerName:            wrongScalerName,
 			MonitoredDeploymentName:    monitoredDeploymentName,
 			ClientName:                 clientName,
@@ -407,6 +477,7 @@ func getTemplateData() (templateData, []Template) {
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
 			{Name: "monitoredDeploymentTemplate", Config: monitoredDeploymentTemplate},
 			{Name: "scaledObjectTemplate", Config: scaledObjectTemplate},
+			{Name: "scaledJobTemplate", Config: scaledJobTemplate},
 			{Name: "clientTemplate", Config: clientTemplate},
 			{Name: "authenticatioNTemplate", Config: authenticationTemplate},
 			{Name: "cloudEventHTTPReceiverTemplate", Config: cloudEventHTTPReceiverTemplate},
@@ -440,7 +511,8 @@ func testScalerMetricValue(t *testing.T) {
 		for _, metric := range metrics {
 			labels := metric.GetLabel()
 			for _, label := range labels {
-				if *label.Name == labelScaledObject && *label.Value == scaledObjectName {
+				if (*label.Name == labelScaledObject && *label.Value == scaledObjectName) ||
+					(*label.Name == labelScaledJob && *label.Value == scaledJobName) {
 					assert.Equal(t, float64(4), *metric.Gauge.Value)
 					found = true
 				}
@@ -484,12 +556,47 @@ func testScaledObjectErrors(t *testing.T, data templateData) {
 	KubectlApplyWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
 }
 
+func testScaledJobErrors(t *testing.T, data templateData) {
+	t.Log("--- testing scaled job errors ---")
+
+	KubectlDeleteWithTemplate(t, data, "scaledJobTemplate", scaledJobTemplate)
+	KubectlApplyWithTemplate(t, data, "wrongScaledJobTemplate", wrongScaledJobTemplate)
+
+	// wait for 2 seconds as pollinginterval is 2
+	time.Sleep(20 * time.Second)
+
+	family := fetchAndParsePrometheusMetrics(t, fmt.Sprintf("curl --insecure %s", kedaOperatorPrometheusURL))
+	if val, ok := family["keda_scaled_job_errors"]; ok {
+		errCounterVal1 := getErrorMetricsValue(val)
+
+		// wait for 2 seconds as pollinginterval is 2
+		time.Sleep(2 * time.Second)
+
+		family = fetchAndParsePrometheusMetrics(t, fmt.Sprintf("curl --insecure %s", kedaOperatorPrometheusURL))
+		if val, ok := family["keda_scaled_job_errors"]; ok {
+			errCounterVal2 := getErrorMetricsValue(val)
+			assert.NotEqual(t, errCounterVal2, float64(0))
+			assert.GreaterOrEqual(t, errCounterVal2, errCounterVal1)
+		} else {
+			t.Errorf("metric not available")
+		}
+	} else {
+		t.Errorf("metric not available")
+	}
+
+	KubectlDeleteWithTemplate(t, data, "wrongScaledJobTemplate", wrongScaledJobTemplate)
+	KubectlApplyWithTemplate(t, data, "scaledJobTemplate", scaledJobTemplate)
+}
+
 func testScalerErrors(t *testing.T, data templateData) {
 	t.Log("--- testing scaler errors ---")
 
 	KubectlDeleteWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
 	time.Sleep(2 * time.Second)
 	KubectlApplyWithTemplate(t, data, "wrongScaledObjectTemplate", wrongScaledObjectTemplate)
+
+	KubectlDeleteWithTemplate(t, data, "scaledJobTemplate", scaledJobTemplate)
+	KubectlApplyWithTemplate(t, data, "wrongScaledJobTemplate", wrongScaledJobTemplate)
 
 	family := fetchAndParsePrometheusMetrics(t, fmt.Sprintf("curl --insecure %s", kedaOperatorPrometheusURL))
 	val, ok := family["keda_scaler_errors"]
@@ -509,6 +616,8 @@ func testScalerErrors(t *testing.T, data templateData) {
 			assert.GreaterOrEqual(t, errCounterVal2, errCounterVal1)
 		}
 	}
+	KubectlDeleteWithTemplate(t, data, "wrongScaledJobTemplate", wrongScaledJobTemplate)
+	KubectlApplyWithTemplate(t, data, "scaledJobTemplate", scaledJobTemplate)
 
 	KubectlDeleteWithTemplate(t, data, "wrongScaledObjectTemplate", wrongScaledObjectTemplate)
 	time.Sleep(2 * time.Second)
@@ -558,6 +667,16 @@ func getErrorMetricsValue(val *prommodel.MetricFamily) float64 {
 			labels := metric.GetLabel()
 			for _, label := range labels {
 				if *label.Name == "scaledObject" && *label.Value == wrongScaledObjectName {
+					return *metric.Counter.Value
+				}
+			}
+		}
+	case "keda_scaled_job_errors":
+		metrics := val.GetMetric()
+		for _, metric := range metrics {
+			labels := metric.GetLabel()
+			for _, label := range labels {
+				if *label.Name == "scaledJob" && *label.Value == wrongScaledJobName {
 					return *metric.Counter.Value
 				}
 			}
@@ -614,7 +733,8 @@ func testScalerMetricLatency(t *testing.T) {
 		for _, metric := range metrics {
 			labels := metric.GetLabel()
 			for _, label := range labels {
-				if *label.Name == labelScaledObject && *label.Value == scaledObjectName {
+				if (*label.Name == labelScaledObject && *label.Value == scaledObjectName) ||
+					(*label.Name == labelScaledJob && *label.Value == scaledJobName) {
 					assert.Equal(t, float64(0), *metric.Gauge.Value)
 					found = true
 				}
@@ -674,7 +794,8 @@ func testScalerActiveMetric(t *testing.T) {
 		for _, metric := range metrics {
 			labels := metric.GetLabel()
 			for _, label := range labels {
-				if *label.Name == labelScaledObject && *label.Value == scaledObjectName {
+				if (*label.Name == labelScaledObject && *label.Value == scaledObjectName) ||
+					(*label.Name == labelScaledJob && *label.Value == scaledJobName) {
 					assert.Equal(t, float64(1), *metric.Gauge.Value)
 					found = true
 				}